działa menu i opcje

+++ Symulacja hodowli robaków - Kopia/Symulacja hodowli robaków/const.h
@@ -1,7 +1,7 @@
 #pragma once
 static const float SCRN_HEIGHT = 720.0f;
 static const float SCRN_WIDTH = 1280.0f;
-static const int TILE_SIZE = 64.0f;
+static const float TILE_SIZE = 64.0f;
 
 enum Face {
 	DOWN, LEFT, UP, RIGHT
